remove legacy code
